requirements.txt
neo4j>=5.0
python-dotenv>=1.0
